int is to small for facebook video upload limit values #181

+++ spring-social-facebook/src/main/java/org/springframework/social/facebook/api/impl/json/VideoUploadLimitsMixin.java
@@ -28,7 +28,7 @@ abstract class VideoUploadLimitsMixin extends FacebookObjectMixin {
 
 	@JsonCreator
 	VideoUploadLimitsMixin(
-			@JsonProperty("length") int length, 
-			@JsonProperty("size") int size) {}
+			@JsonProperty("length") long length, 
+			@JsonProperty("size") long size) {}
 	
 }
